organize the code for CPI, PI and LOCO in 3 modules + one base class in a base_perturbation.py module. Rename the module that contains the function version
